Move community folder outside of modals
