+++ src/modules/chat/HomePage/MyChatsContent.tsx
@@ -2,8 +2,8 @@ import NoResultImage from '@/assets/graphics/no-result.png'
 import Button from '@/components/Button'
 import ChatPreviewList from '@/components/chats/ChatPreviewList'
 import ChatPreviewSkeleton from '@/components/chats/ChatPreviewSkeleton'
+import NewCommunityModal from '@/components/community/NewCommunityModal'
 import Container from '@/components/Container'
-import NewCommunityModal from '@/components/modals/community/NewCommunityModal'
 import { COMMUNITY_CHAT_HUB_ID } from '@/constants/hubs'
 import { getPostQuery } from '@/services/api/query'
 import {
